% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AADVMTModel_df}
\alias{AADVMTModel_df}
\title{list-column data frame (tibble) with estimated model objects for household AADVMT model}
\format{The main data frame \code{AADVMTModel_df} has 2 rows and 3 variables:
\describe{
  \item{metro}{metro or non-metro segment of the AADVMT model}
  \item{model}{the model objects stored in list-column of the data frame}
  \item{post_func}{the function for post-processing model predictions}
  }}
\usage{
AADVMTModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household AADVMT model
Estimated with the script in data-raw/AADVMTModel_df.R
}
\examples{
str(AADVMTModel_df)
head(AADVMTModel_df)
summary(AADVMTModel_df)

}
\keyword{datasets}
