% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BikePMTModel_df}
\alias{BikePMTModel_df}
\title{list-column data frame (tibble) with estimated model objects for household Biking PMT model}
\format{The main data frame \code{BikePMTModel_df} has 2 rows and 3 variables:
\describe{
  \item{metro}{metro or non-metro segment of the Biking PMT model}
  \item{model}{the model objects stored in list-column of the data frame}
  \item{post_func}{the function for post-processing model predictions}
  }}
\usage{
BikePMTModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household Biking PMT model
Estimated with the script in data-raw/BikePMTModel_df.R
}
\examples{
str(BikePMTModel_df)
head(BikePMTModel_df)
summary(BikePMTModel_df)

}
\keyword{datasets}
