% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetHouseholdDvmtDL.R
\name{BudgetHouseholdDvmtDL}
\alias{BudgetHouseholdDvmtDL}
\title{Calculate household DVMT and adjusts based on budget and calculates trips}
\usage{
BudgetHouseholdDvmtDL(L, M)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}

\item{M}{A list the module functions of modules called by this module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{BudgetHouseholdDvmtDL} calculate the average household DVMT, adjusts it
to fit within the household budget, and calculates vehicle trips and
alternative mode trips.
}
\details{
This function calculates the average household DVMT and adjusts it to fit
within the household budget. Using the adjusted DVMT and other
characteristics, the module also calculates household vehicle trips and
alternative mode trips. It also recalculates urban and rural DVMT.
}
