% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetHouseholdDvmtDL.R
\docType{data}
\name{BudgetHouseholdDvmtDLSpecifications}
\alias{BudgetHouseholdDvmtDLSpecifications}
\title{Specifications list for BudgetHouseholdDvmtDL module}
\format{
A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
 \item{Call}{modules that are called by the module}
}
}
\source{
BudgetHouseholdDvmtDL.R script.
}
\usage{
BudgetHouseholdDvmtDLSpecifications
}
\description{
A list containing specifications for the BudgetHouseholdDvmtDL module.
}
\keyword{datasets}
