% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAltModeTrips.R
\name{CalculateAltModeTrips}
\alias{CalculateAltModeTrips}
\title{Main module function to calculate alternative mode trips}
\usage{
CalculateAltModeTrips(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateAltModeTrips} Function first predicts PMT for Walk, Bike and Transit modes 
for each household in the households dataset using independent variables including household 
characteristics and 5D built environment variables.
}
\details{
This function predicts PMT for alterantive modes for each hosuehold in the model region where
each household is assigned a PMT value for each mdoe. The model objects as a part of the
inputs are stored in data frame with two columns: a column for segmentation
(e.g., metro, non-metro) and a 'model' column for model object (list-column
data structure). The function "nests" the households data frame into a
list-column data frame by segments and applies the generic predict() function
for each segment to predict PMT for each household. This is done for all modes and 
vectors of HhId and PMT values are produced and stored in the
"Household" table.

If this table does not exist, the function calculates a LENGTH value for
the table and returns that as well. The framework uses this information to
initialize the Households table. The function also computes the maximum
numbers of characters in the HhId and Azone datasets and assigns these to a
SIZE vector. This is necessary so that the framework can initialize these
datasets in the datastore. All the results are returned in a list.

The second part of function predicts the the daily average trip length and 
annual averae number of trips generated by each alternative modes in households
and return the values to the datastore
}
