% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPredictions.R
\name{combine_preds}
\alias{combine_preds}
\title{Used to combine predictions from multiple models: 
 for example combine GreenSTEP zero dvmt glm model and dvmt lm model
 Example (a clumsy way to square mpg):
 mtcars %>% 
   nest(-cyl) %>%
   crossing(tibble(x=c(1, 2))) %>%
   mutate(mpg=map(data, "mpg")) %>% 
   group_by(cyl) %>%
   summarise(mpg1 = list(mpg[[1]]),
             mpg2 = combine_preds(mpg)) %>%
   mutate(test=map2_lgl(mpg1, mpg2, ~all(.y==.x^2)))}
\usage{
combine_preds(preds_ls, func = `*`, init = 1, ...)
}
\description{
Used to combine predictions from multiple models: 
 for example combine GreenSTEP zero dvmt glm model and dvmt lm model
 Example (a clumsy way to square mpg):
 mtcars %>% 
   nest(-cyl) %>%
   crossing(tibble(x=c(1, 2))) %>%
   mutate(mpg=map(data, "mpg")) %>% 
   group_by(cyl) %>%
   summarise(mpg1 = list(mpg[[1]]),
             mpg2 = combine_preds(mpg)) %>%
   mutate(test=map2_lgl(mpg1, mpg2, ~all(.y==.x^2)))
}
