% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPredictions.R
\name{CombinePreds}
\alias{CombinePreds}
\title{Combine predictions from multiple models: for example combine GreenSTEP zero
dvmt glm model and dvmt lm model As an example (a clumsy way to square mpg):}
\usage{
CombinePreds(preds_ls, func = `*`, init = 1, ...)
}
\arguments{
\item{preds_ls}{A list of prediction vectors for each modeling steps}

\item{func}{function used to combine results from modeling steps}

\item{init}{initial value, pass to the `.init` of `reduce()`}

\item{...}{additional argument passed to `reduce`}
}
\value{
a list with the combined predictions
}
\description{
Combine predictions from multiple models: for example combine GreenSTEP zero
dvmt glm model and dvmt lm model As an example (a clumsy way to square mpg):
}
\examples{
 library(tidyverse)
 mtcars \%>\%
   nest(-cyl) \%>\%
   crossing(tibble(x=c(1, 2))) \%>\%
   mutate(mpg=map(data, "mpg")) \%>\%
   group_by(cyl) \%>\%
   summarise(mpg1 = list(mpg[[1]]),
             mpg2 = CombinePreds(mpg)) \%>\%
   mutate(test=map2_lgl(mpg1, mpg2, ~all(.y==.x^2)))
}
