% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPredictions.R
\name{DoPredictions}
\alias{DoPredictions}
\title{Helper functions for running predictions}
\usage{
DoPredictions(Model_df, Dataset_df, dataset_name, id_name, y_name,
  SegmentCol_vc = NULL, combine_preds = FALSE)
}
\arguments{
\item{Model_df}{A list-column data_frame containing a 'model' list-column,
column(s) used for segmenting data, and, optionally, a 'step' column and a
'post_func' for multi-step models and post-processing function for
predictions.}

\item{Dataset_df}{A data frame providing independent variables and
segmentation column for predictions included in the specification of the
model objects.}

\item{dataset_name}{A character string for name of the Dataset_df}

\item{id_name}{A character string for the id column of Dataset_df}

\item{y_name}{A character string for name of the outcome variable}

\item{SegmentCol_vc}{A vector for columns used for segmentation; if NULL
assuming columns other than c("model", "step", "post_func", "bias.adj") in
Model_df is used for segmentation}

\item{combine_preds}{A logical flag indicating whether to combine predictions
for multi-step models (default FALSE)}
}
\value{
A list containing the components specified in the Set specifications
  for the module along with: LENGTH: A named integer vector having a single
  named element, "Household", which identifies the length (number of rows) of
  the Household table to be created in the datastore. SIZE: A named integer
  vector having two elements. The first element, "Azone", identifies the size
  of the longest Azone name. The second element, "HhId", identifies the size
  of the longest HhId.
}
\description{
run predictions on Dataset_df with model objects in Model_df
}
\details{
\code{DoPredictions} predicts outcomes (y) for each observation in the
Dataset_df dataset using independent variables included in Dataset_df and
model objects saved in list-column data_frame Model_df.
}
