% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPredictions.R
\name{.y(.x)}
\alias{.y(.x)}
\title{internal function that handles pass a list column of a data frame to another
list column as argument.}
\usage{
`.y(.x)`(.x, .y, ...)
}
\arguments{
\item{.x}{the data argument for function call `.y(.x)`}

\item{.y}{the function to be used for call `.y(.x)`}

\item{...}{other arguments passing to `.y()` along with `.x`}
}
\value{
a list with result from call `.y(.x, ...)`
}
\description{
internal function that handles pass a list column of a data frame to another
list column as argument.
}
\examples{
 library(tidyr)
 library(dplyr)
 library(purrr)
 lcdf <- mtcars \%>\% nest(-cyl)
 lcdf$fmla <- list(~lm(mpg~wt, data=.x))
 lcdf <- lcdf \%>\% mutate(fit=map2(data, fmla, VETravelDemandMM:::`.y(.x)`))
 lcdf$fit \%>\% map(summary)

}
