% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDrivers.R
\docType{data}
\name{PredictDriversSpecifications}
\alias{PredictDriversSpecifications}
\title{Specifications list for PredictDrivers module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\usage{
PredictDriversSpecifications
}
\description{
A list containing specifications for the PredictDrivers module.
}
\keyword{datasets}
