% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictTransitPMT.R
\name{PredictTransitPMT}
\alias{PredictTransitPMT}
\title{Main module function}
\usage{
PredictTransitPMT(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module along with:
LENGTH: A named integer vector having a single named element, "Household",
which identifies the length (number of rows) of the Household table to be
created in the datastore.
SIZE: A named integer vector having two elements. The first element, "Azone",
identifies the size of the longest Azone name. The second element, "HhId",
identifies the size of the longest HhId.
}
\description{
\code{PredictTransitPMT} predicts TransitPMT for each household in the households
dataset using independent variables including household characteristics
and 5D built environment variables.
}
\details{
This function predicts TransitPMT for each hosuehold in the model region where
each household is assigned an TransitPMT. The model objects as a part of the
inputs are stored in data frame with two columns: a column for segmentation
(e.g., metro, non-metro) and a 'model' column for model object (list-column
data structure). The function "nests" the households data frame into a
list-column data frame by segments and applies the generic predict() function
for each segment to predict TransitPMT for each household. The vectors of HhId
and TransitPMT produced by the PredictTransitPMT function are to be stored in the
"Household" table.

If this table does not exist, the function calculates a LENGTH value for
the table and returns that as well. The framework uses this information to
initialize the Households table. The function also computes the maximum
numbers of characters in the HhId and Azone datasets and assigns these to a
SIZE vector. This is necessary so that the framework can initialize these
datasets in the datastore. All the results are returned in a list.
}
