% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVehicles.R
\name{PredictVehicles}
\alias{PredictVehicles}
\title{Main module function to predict Vehicles for each household}
\usage{
PredictVehicles(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictVehicles} predicts the number of Vehicles for each
household and tallies the total number of Vehicles for each household.
}
\details{
This function predicts the number of Vehicles for each household. It uses
}
