% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVehicles.R
\docType{data}
\name{PredictVehiclesSpecifications}
\alias{PredictVehiclesSpecifications}
\title{Specifications list for PredictVehicles module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\usage{
PredictVehiclesSpecifications
}
\description{
A list containing specifications for the PredictVehicles module.
}
\keyword{datasets}
