% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictWalkTFL.R
\name{PredictWalkTFL}
\alias{PredictWalkTFL}
\title{Main module function}
\usage{
PredictWalkTFL(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module along with:
LENGTH: A named integer vector having a single named element, "Household",
which identifies the length (number of rows) of the Household table to be
created in the datastore.
SIZE: A named integer vector having two elements. The first element, "Azone",
identifies the size of the longest Azone name. The second element, "HhId",
identifies the size of the longest HhId.
}
\description{
\code{PredictWalkTFL} predicts person miles traveled by Walk for each
household in the households dataset using independent variables including
household characteristics and 5D built environment variables.
}
\details{
This function predicts person miles traveled by Walk for each hosuehold in the model
region. The model objects as a part of the inputs are stored in data frame
with two columns: a column for segmentation (e.g., metro, non-metro) and a
'model' column for model object (list-column data structure). The function
"nests" the households data frame into a list-column data frame by segments
and applies the generic predict() function for each segment to predict trip
frequency and length for each household. The vectors of HhId and
PMTWalk produced by the PredictWalkTFL function are to be stored in
the "Household" table.
}
