% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TransitPMTModel_df}
\alias{TransitPMTModel_df}
\title{list-column data frame (tibble) with estimated model objects for household Walking PMT model}
\format{The main data frame \code{TransitPMTModel} has 2 rows and 3 variables:
\describe{
  \item{metro}{metro or non-metro segment of the Transit PMT model}
  \item{model}{the model objects stored in list-column of the data frame}
  \item{post_func}{the function for post-processing model predictions}
  }}
\usage{
TransitPMTModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household Transit PMT model
Estimated with the script in data-raw/TransitPMTModel.R
}
\examples{
str(TransitPMTModel_df)
head(TransitPMTModel_df)
summary(TransitPMTModel_df)

}
\keyword{datasets}
