% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TransitTFLModel_df}
\alias{TransitTFLModel_df}
\title{list-column data frame (tibble) with estimated model objects for household Transit TFL model}
\format{
The main data frame \code{TransitTFLModel} has 2 rows and 3 variables:
\describe{
  \item{metro}{metro or non-metro segment of the Transit TFL model}
  \item{model}{the model objects stored in list-column of the data frame}
  \item{post_func}{the function for post-processing model predictions}
  }
}
\usage{
TransitTFLModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household Transit TFL model
Estimated with the script in data-raw/TransitTFLModel.R
}
\examples{
str(TransitTFLModel_df)
head(TransitTFLModel_df)
summary(TransitTFLModel_df)

}
\keyword{datasets}
