% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VehiclesModel_df}
\alias{VehiclesModel_df}
\title{list-column data frame (tibble) with estimated model objects for household vehicles model}
\format{The main data frame \code{VehiclesModel} has 1 rows and 2 variables:
\describe{
  \item{model}{the model objects stored in list-column of the data frames}
  \item{post_func}{the function for post-processing model predictions}
  }}
\usage{
VehiclesModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household vehicles model
Estimated with the script in data-raw/VehiclesModel.R
}
\examples{
str(VehiclesModel_df)
head(VehiclesModel_df)
summary(VehiclesModel_df)

}
\keyword{datasets}
