% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VETravelDemandMM.R
\docType{package}
\name{VETravelDemandMM}
\alias{VETravelDemandMM}
\alias{VETravelDemandMM-package}
\title{\code{VETravelDemandMM} package}
\description{
Simulate Multi-Modal Travel Demand for Households
}
\details{
See the README on
\href{https://github.com/gregorbj/VisionEval/sources/modules/VETravelDemandMM}{GitHub}
}
