% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WalkPMTModel_df}
\alias{WalkPMTModel_df}
\title{list-column data frame (tibble) with estimated model objects for household Walking PMT model}
\format{The main data frame \code{WalkPMTModel} has 2 rows and 3 variables:
\describe{
  \item{metro}{metro or non-metro segment of the Walking PMT model}
  \item{model}{the model objects stored in list-column of the data frames}
  \item{post_func}{the function for post-processing model predictions}
  }}
\usage{
WalkPMTModel_df
}
\description{
list-column data frame (tibble) with estimated model objects for household Walking PMT model
Estimated with the script in data-raw/WalkPMTModel.R
}
\examples{
str(WalkPMTModel_df)
head(WalkPMTModel_df)
summary(WalkPMTModel_df)

}
\keyword{datasets}
