% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDrivers.R
\name{PredictDrivers}
\alias{PredictDrivers}
\title{Main module function to predict Drivers for each household}
\usage{
PredictDrivers(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictDrivers} predicts the number of Drivers for each
household and tallies the total number of Drivers for each household. It uses
the model specifications in the model object in data/DriversModel_df.rda.
}
